// src/app/shared/interfaces/mat-columns.ts
export interface MatColumns {
    /**
     * @ var title String Titre de la colonne dans le tableau
     */
    title: String;

    /**
     * @ var always Boolean Détermine si la colonne doit toujours être affiché
     */
    always: Boolean;

    /**
     * @ var value String: valeur utilisées dans la liste des colonnes du tableau 
     */
    value: String;

    /**
     * @var isDisplayed Boolean Vrai si la colonne est affichée
     */
    isDisplayed: Boolean;
}